Change message in arrivals when the stop name is found
Summary: Fix T1368

Test Plan:
Check stop 17369, or a stop that is not serviced at the moment (like late at night). It suffices to have a name in the DB (therefore, present of the map)

If the message says "no arrivals found" instead of "no stop with this name", it works.

Reviewers: valerio.bozzolan

Reviewed By: valerio.bozzolan

Subscribers: valerio.bozzolan, #libre_busto

Tags: #libre_busto

Maniphest Tasks: T1368

Differential Revision: https://gitpull.it/D208